hare_lab_bot/requirements.txt
aiogram==2.13
aioschedule~=0.5.2
PyYAML~=6.0
pendulum~=2.1.2
requests~=2.25.1
pydantic~=1.8.2